nrunner: split spawners to own package
Spawners are not being used by the current nrunner module. To improve
readability and future grow this small change will organize the spawners into
modules.
